Makes Dexalin, Cryo chems, and some advanced meds more sensible
Vermicetol, formerly requiring 0.1u of plasma, instead requires 0.5u each of cryoxadone and clonexadone.

Cordradaxon, Peridaxon, and Respirodaxon all require 0.1u of clonexadone, of which they are each a derivative chemical

Carthatoline now requires 0.1u of toxin instead of plasma and results in 3 units instead of 2, making dylovene/carbon injections a viable way to treat the effects of small amounts of toxin (such as those injected by cuts from digging through the trash piles) in the blood without dialysis.

Varceptol now requires carthatoline instead of dylovene.

Dexalin consumes iron (simulating hemoglobin) instead of plasma, and now requires 5u as a catalyst.

Dexalin Plus will not mix in temperatures exceeding 180 degrees Kelvin

Cryoxadone now requires dexalin plus instead of normal dexalin, and will not mix in temperatures exceeding 90 degrees Kelvin.

Clonexadone now requires radium instead of plasma, and will not mix in temperatures exceeding 45 degrees Kelvin.

All in all, these chemicals are rarely ever mixed due to the consumption of plasma imposed on them for the purposes of 'balance' by Baystation back in the day, in order to make the game 'harder' by combining a catalyst requirement of 5u with consumption of that catalyst. This PR seeks to remedy the situation by making the relevant chemicals cheaper to make, at the cost of being made with a more involved process.