Hard nerf to render HP
Buffing renders to ridiculous levels solely because of ONE PLAYER doing ridiculous shit after equally ridiculous levels of preparation is not good gameplay balance, as it punishes everyone for the actions of one or two autists. These things are so goddamn fast already that you will be forced into a melee encounter with them and they hit so hard that melee is entirely unviable in the first place. THese things should not be taking out a party of three well-prepared people with hard-hitting guns. Vigil MacLeod is not a valid reason for over-buffing renders. It still has 500 HP with this change, which is plenty enough to make soloing them very fucking difficult if not impossible with all their other advantages.