Pool tiles and water changes
About The Pull Request
I've asked around and people seem to be receptive to the idea of having pool tiles and effects, as well as water overlays. The overlays are effects that can be applied to any tile via map edits. All water tiles, along with the pool tiles now have water footstep sounds when someone walks over them. More effects and such to come in the future, maybe.

Changelog
🆑
add: Pool tiles that play water sounds when walked over, have a splash effect when clicked on by a person with empty hands.
Add: water overlay effects that can be added to any tile via map edits.
Add: water1-4.ogg to the sound files to play for footsteps.
change: foot_sound can now accept "water" to get the desired splooshes from wet tiles if desired.
Add: Many icons for use in the pool and various effects that can be used in the future.
/🆑