Scrap weapons!
A laser SMG remake has been made for the texan smg, rebalanced slightly off the ERIS version as the ERIS version was stupidly powerful (deals 80 damage if all shots hit). Also rebalanced the scrap rifle to be slightly stronger, it fires a bit faster and now deals more damage but is still inferior to the rail rifle. Both can now be found in scrap piles with a high weight.