Organ Refactor (For Carrion)
This push fixes a lot of existing issues with Carrion and gets its (nearly) running perfectly. Theirs a runtime issue with updating the mon sprite when a spider is implanted, it runtimes and the mob_implant icon doesn't show. It makes examining someone impossible until the spider is removed. I also think that an issue with surgery cut opens are not displaying either due to a related bug, pushing this to check point my work before I try some experimental ports.