Tables can be used as descent ladders.
If you're on the floor above an area with open space and a table is there you can click drag to descend down their safely at the same speed one would climbing onto a table. Duration is controlled by vigilance. The trip is one way only and can be blocked if the table is full/occupied.