Better Litanies
The prayer of reunion now allows you to select a cruciform by path, installing it grants you your pathway litanies, and the upgrade kit litany is cleaner and now requires an altar to use properly. Cruciforms in a person can now be repaired and reactivated with a new litany. Pain relief litanies now heal halloss in addition to agony and a painkiller effect as otherwise the effect they did was basically unnoticable. Perfect self now last 15 minutes.