AdminAddPerk1.2
Changed it so that you now get a list, list is retrieved from an available perk pool when adding and from the creature's perk pool when removing.

If there is no perk to remove then you are told that there's none to remove and we return before attempting to remove it in this version.

Tested with adding and removing, both as an admin verb and through view variables.