Admin Add and Remove Perks (#484)
* Admin Add and Remove Perks Version 1.

Perk name is inputted manually, has to be the tailend of the path from https://github.com/sojourn-13/sojourn-station/blob/aca66e8cbbc9c32f48237728d838af8e87623db9/code/__DEFINES/perks.dm

So sharp_mind would be for the Sharp Mind perk.

Might implement a list later, but this would require implementation with perks.dm to dynamically react when new perks are added. It's an admin command, doesn't need too much finesse at the moment.

* AdminAddPerk1.2

Changed it so that you now get a list, list is retrieved from an available perk pool when adding and from the creature's perk pool when removing.

If there is no perk to remove then you are told that there's none to remove and we return before attempting to remove it in this version.

Tested with adding and removing, both as an admin verb and through view variables.

* AdminAddPerk1.3

Added a check if the mob is deleted to cut the operation short. Added a way to not go forward with the operation and instead cancel out.