Dismemberment Fix
Organs normally have their 'owner' variable set to null when removed, but external organs re-write the removed_mob and forget that part.

This causes an issue when a limb is destroyed, causing the owner to no longer be able to use that limb slot for the round even if they got a new limb.

This fixes this problem. Tested on my own server.