Disarm Intent Beaker Pour [glass.dm update]
This allows players while on Disarm intent to pour beakers into another beaker/chem bottle/vial without splashing or down into a sink.  This is a quicker way if you want to just use the entire bottle instead of having to right-click context menu and select the max amount everytime.