Adjusts Code to Requests
Brings back the Augmentation Spinny Thingie.

getFlatIcon() is still not called by updatePreviewIcons(), so we are saving processing, but if in the case that a player specifically spins their model to face East on the Augmentation screen, an Icon will be spontaneously generated, and immediately cached to integrate with the Augmentation Screen's code.