Cossack Coats
Three different coats, one that reflects the frontier Jaegers, another reflects how Jaeger are a folk of celebrations and community.
The last piece of armor is locked off to Marshals and Blackshield, it's as strong as a generic armor vest.  It's made to fit with with the blackshield color scheme, while the cuffs and color fit the Marshal color scheme.