Hunter Gear Update 1
Give some more love to the Hunter's Lodge in the form of 4 new exclusive items:
-New boots, come with two different styles that you can swap between - Tahca and mukhwa.
-New gloves
-New weapon! The Hunting Claw
-New sheath, made specifically for the Hunting Claw
Recipes for all of them, exclusive to Hunters
They are slightly upgraded variants of anything that can be spawned or colony made and come with a base price tag of 200 credits (except the Claw and sheath, which you shouldn't trade really) to give the Hunters something to actually trade and export apart from animal fangs and cakes.
No, you don't spawn with them, no, not even the Huntmaster. Fuck you, craft them.
Credits:
HUGE shoutout to Zienite for holding my hand through the entire process and patiently answering a hundred questions
To me, for making the cool sprites