Time Tracking 2.0
We now enable the last part of time tracking, namely the locking. You can check how much time you've played in each job under IC. Currently need to have played 20 hours in that department to be able to play the head of staff, or 20 in the council in general to play Premier and Steward.

Tested in my own test environment (a lot of times). Things should work as expected.