Code and Mapping
This is a long list of stuff that I spent like four or so hours doing, doing a quick summary.
Code:
Moved around some of the magazines on other disks to be more in order/proper.
Added a few more mags and ammo boxes to the Blackshield sovereign disk to make it more 'Sovereign', they will still need to go to Guild or Marshals for more ammo/different ammo.

Mapping:
Added a CURSED bathroom to the Blackshield Barracks.
Made the other way to tunnel into the colony just as dangerous as the other tunnel by the crematorium.
Littered the colony grounds between perimeter and colony with more trapdoor burrows
Moved organ freezer from Shield to Corpsman triage, removed box of latex gloves, replaced all five first aid kits with a singular Blackshield siege kit.
Slapped in a box of handcuffs into command armory for steward.
Gave BS their box of crayons.
Created a single way for someone to mine into BS on the staircase zone.