Genetics fixes
Removed verbs for interacting with the cloner through object interactions. Now you have to operate it via the console exclusively.

Fixed Icon updating for the bidon can, so the correct anchoring sprites show.

Fixed an issue causing a majority of the cloning recipes to not function correctly. Go make some chocolate milk cows!