genetics tweaks
Fixes Pulpers so shoving in robots and such don't spawn sample plates (Thanks Hayley)

Moves the genetics studio design disk to the CRO's office.

Adds a new genetics studio resupply disk to the Genetics lab. It has just the purger and mutagenic injector on it.

Adds sprites for both disks (Credit to @TripleZ#1242 )