HOTFIX! CHECK BEFORE MERGE PLEASE
moves the var (show_title) to the human.dm file. This means the card/id file is no longer modified by this PR.
Creates the verb Toggle Title for the IC tab and has it change on the person instead of card. This means dropping or losing your ID card wont cause the game to go OH FUCK.