Adds New blood chems, Fixes Sanguinum OD
Fixes sanguinum overdose. Its now the same as any other overdose. No more putting hundreds of units into you without consiquence.
Adds two chems, nosfernium and viroputine. I have more plans for viroputine down the road.
Nosfernium is a much stronger blood restore with toxin damage and bad withdrawels.
Viroputine is a unique chem in that you feed it blood and chems with nosfernium to make more. (This is why the innitial making of nosfernium is so low)