Fixed rocket damage
The previous damage of the rocket was 75, this is less than an AMR, and less than a shunted reductor. One of the rarest and heaviest weapons in the game, not only that, but designed for anti-tank duties, did less damage than a weapon the guild can manufacture anywhere, and have a shitty cell and cheap soteria mod put in.