Dye_Vat
Base functionallity added to slime dye vat. Its capable of taking empty and full slime jars to get uses that it uses up to dye clothes. You set the color with crayons. Certain crayons are currently not showing the correct color. Namely, Orange and Rainbow. Rainbow slimes are able to be made with the dye vat by attempting to dye something after changing the color 5 plus times with the current color being rainbow when you attempt to dye.