Update _Nadezhda_Colony_Omni_STU.dmm
LITERALLY 1984. Maps in the E.Y.E Surveillance pod to the Marshals tactical storage, a piece of useful equipment with unusual demands.
To use it, simply put on the provided equipment in the room, turn on the provided internals and climb in before clicking the pod.
It functions similarly to the AIs eye, allowing you to pan through the camera network in a highly efficient fashion. Credit goes to @KIROV#6255 for the actual code of the thing!