Knight rigsuit cost nerf (#3248)
Material costs of the knight rigsuit doubled for biomatter, gold, silver. Material cost of plasteel raised to 100. While these numbers may seem high it's worth noting that printed from a bioprinter the final value will only be 60% of these numbers. This means that the new, total cost for the suit is lower than initially seems.

Co-authored-by: cdb-is-not-good <jaeger_avro@hotmailc.om>