LITERALLY 1984: 2, cyberpunk boogaloo (#3264)
* 1984 2

Slight map adjustment! Evidence storage has been moved to the top right corner of the 'investigation block' in the central Marshals, in its place is the E.Y.E Pods storage room, locked to Ranger access(All Marshals will have access thusly.)
To use it, simply put on the provided equipment in the room, turn on the provided internals and climb in before clicking the pod.
It functions similarly to the AIs eye, allowing you to pan through the camera network in a highly efficient fashion. Credit goes to @Kirov#6255 for the actual code of the thing!

* actually fixes the stupid fucking tint button

* Update _Nadezhda_Colony_Omni_STU.dmm

Co-authored-by: cdb-is-not-good <jaeger_avro@hotmailc.om>