Mask Voice Transmitters (#3352)
There is a difference between muffled and having a different voice now. Added the ability to alter the transmitters in gas masks to turn you unknown. This is done with a screwdriver on ANY gas mask.
Adds in the ability to put voice recognizers from the assembly system onto any mask to give it the same effect as the gas masks. This won't hide your face if the mask dosn't. BUT IT DOES WORK ON SNORKALS! Be warned it won't come back off once installed.