By assortedbeads
balance:
Rebalanced the damage of the old spear so it properly references our damage define.
Arranged the new spears in an increasing tier list, making sure none are as strong as the current NT spears.

imageadd:
Added icons for the 4 types of spears, which show up inhand, wielded and on back and suit storage

code:
Added recipes for crafting all the spears

code:
Put the throw speed of the new spears and the already added throwing knife in line with current ranged options.
Gave the various spears style damage when thrown according to their tier, change in line with Pilum.