Basic Chemist Offers, Combination Driver Slash (With a hint of culinary tariffs)
Sets non-default prices for meat, tofu, eggs, flour, and milk for Vermouth station.

Cad station has gotten a set of basic medicine bottles as offers, to give something chemists to do quickly and easily. Or the church/garden that knows how to work with plants.

Gives Dylovene a non-default price for the SI station.

Decreases the value of both normal and greyson combination drills, and ups the required materials to print both.