Small EMP rocket nerf
The EMP rocket is essentially a specialised ammunition type made against mechas or generally, things that are vulnerable to EMPs. It's understandable that the EMP warhead will likely take the place of most sorts of kinetic penetrators, HE filler and so on, in order to make it able to do its role. So I nerfed the penetration value of the rocket which made no sense, now and some damage, EMP aside, it's just essentially not gonna hurt you as much, to not make it as unbalanced. Just buy regular rockets against people now.