Introduces a new hog mob for service
Currently, there is only one pig in the garden and there's no easy way to get more. In the spirit of giving the chef broader access to pork without copping out and just making it print from the biogenerator or making the existing pig reproduce asexually like the cerberus does, I created a genetic combination of the two as a more lore-friendly means of accomplishing the stated goal.