small language key bugfix.
Fixes a small bug where some languages used the same keys to work.
MarQua speak is now ,q for mar'Qua
Narad pidgin is now ,n for naramad
Cult common is now C for cult
Occult is now Occult Projection to make clear that it is not just a spoken language, key is now ,z for aztral projection or smth.
Also moves technical cant to ,t for the sake of consistency with most other languages as well as changing its shorthand to TC.

Testing procedure: didn't.