fixes  (#4494)
* Naramad Language

New language, Narad Pidgin.
"A crude pidgin tongue, not created by it's supposedly native speakers, but by their foreign overlords. Enforced as the only legal native language of the Naramadi people by the Federation as a means to unify the planet and make administrative work easier, while allowing them to keep a semblance of native culture."

All credit goes to Yanniert for lore + language details. Tested and working as far as i can tell.

* Update code/modules/mob/language/outsider.dm

* small language key bugfix.

Fixes a small bug where some languages used the same keys to work.
MarQua speak is now ,q for mar'Qua
Narad pidgin is now ,n for naramad
Cult common is now C for cult
Occult is now Occult Projection to make clear that it is not just a spoken language, key is now ,z for aztral projection or smth.
Also moves technical cant to ,t for the sake of consistency with most other languages as well as changing its shorthand to TC.

Testing procedure: didn't.

* Makes it actually compile, oops lol <:3c

---------

Co-authored-by: Trilbyspaceclone <30435998+Trilbyspaceclone@users.noreply.github.com>