Allows outsiders to be robots + have core implants
With the general thrust of removing the 'dante must die' difficulty of outsiders, it seems apt to allow them to have access to synths and core implants, particularly because the balance benefit of restricting these to colony only seems offset by the loss to roleplay. EG, in the past we've had wayward church members arrive, we've had opifex who reasonably would have nanogates of their own arrive, we've had synths who are only recently arrived to the planet(from their perspective) arrive. By allowing outsiders once again to be unrestricted, we allow greater forms of roleplay. If outsiders become an issue of doing others jobs, taking over the jungle, etc- then it is perhaps better to handle it from an OOC perspective per our powergaming rules.

Particularly regarding dept synths, those restrictions may be wise to keep given that they ARE departmentally owned- but I would still argue for keeping the restrictions all gone for the above reasons, and adjusting as needed after seeing in play.