Ammo fab access
Removes the perk check on the bullet fab.

Why? The bullet fab is a great tool for security as a whole, but the fab being locked was something originally done because people would frequently break into the guild to use it when they weren't around where they'd likely suffer 0 consequence due to the guild being a low sec area.
the fab is not in the guild anymore. It's in the Marshals. Breaking into the armory is a capital crime that carries extremely severe consequences. This change will additionally allow those with access(Sgt, CO, Premier, Steward) to use the device in the event of emergencies where a WO/Spec might not necessarily be available.