Update belt.dm
Small change to allow sec-belts to hold stun batons like they used to. The ability to hold stun batons was lost ages ago when we switched from obj/item/weapon to obj/item/tool for batons so they'd be more in line with the rest of our tools.

testing: probably works :)