Unfucks a few sprites, buffs marshals,.
Adds in missing sprites for the greyson guns so they won't go invisible when wielded despite being visible when in hand.

also fixes the MArshal thermo NVGs to actually work.

Reverts the change to the COs long-arm, giving it the original sprite. The blackguard sprite is notably inferior and frankly, doesn't look very good period.

additionally adds the missing TK frame to the Marshal gun disk, as well as the omni design that I'm not sure why it wasnt present.

make the heroic craftable with said Takeshi frame and gives it proper pathing and all.

Adjusts some of the prices on the Marshal parts disk, mostly lowering them down to 1 other than the AMR, which has been raised to five points. This is intended as a step between just getting rid of the points cost for the parts disk which was suggested, perhaps this will be sufficient.