Chemistry For the Outlander
For too long, outsiders have been kinda shafted when it comes to medical needs. This gives them access to a chemistry dispenser along with prospectors if they so wish at the void wolf medical facility. These were removed so everyone had to rely on soteria for their medicines, but that's not really an option for outsiders.