Excel armor and mob tweaks
Changed Excel void suit to reflect it's description and make it less of a juggernaut suit made by the gods.
(Requires additional material to craft /&/ Better energy protection, less melee and rad with average Balistic; with added weight as it's a pretty beefy bomb suit for a full void suit)

Seems the experience i've seen is that even a lone Excel agent can 1v5 the colony with this armor, although it's exspected that the excel agent is to be fighting an uphill battle. This will hopefully make them less reliant on going full "nuke-ops war mode" every time they play. Y'all got traps and other fun toys to play with before choosing a gun. After all a dead colonist is a terrible slave for implants of the glorious revolution!.

Reduced excel slave meat drop from 4 to 1 to reduce item lag, and made all mobs carry 3 mags worth of ammo (the ppsh for some reason only had 2 mags)