Small lil bandaid.
Fun fact; Prosthetics did NOT follow the usual code that would cause a limb to regenerate its damage_mask after being fixed(say, with a welder.).

This pr does *not* fix this underlying issue, it does however provide a band-aid to resolve the symptoms of said problem via a forced icon update whenever using a tool with the welding quality, or a cable length to repair a synth.