Fixes dress selection. (#4721)
* Fixes dress selection.

Fixes a small error in the loadout list that prevented selecting from the wide range of Dresses we have had for like. A year. Not sure who broke this.

* Motherfuckers.

Fixes a couple long-time loadout bugs. Dresses are now unified under 1 heading, and now properly work. There's a few duplicates but whatever, it's fine.

Also fixes the issue with the BS fatigues having a one null sprite selectable.