Further tweaks
Baroque is now not stupid. AP down from 250 to 50. Supereffective mult of 6, total damage is roughly the same if not a little lower but ONLY against /mob/living(other than humans.).

Also no longer embeds because why would it? It's a high power hitscan that pierces walls, it goes RIGHT through you.