Buffs marshals for Pneumo. Nerfs marshals for Meeeeeee
Adds the wind(ostwind + Nordwind) frame to the Marshal disk.

Raises the price of the wind frame to include osmium owing to the high quality of the weapons it produces- ditto for the omni frame. In the future I may want to find an alternative resource for these higher quality weapon frames