Visitor update. (#4761)
* Visitor update.

Visitor role is back, however this time it is not simply an alt-title for colonists. They spawn with a 0 access , 'official' guest pass that shows they have the right to be there. Mind you, they are STILL visitors they are STILL subject to all the usual rules for visitors. any crimes they commit can result in ejection and they generally should stick to main areas - but this should help the longtime issue of new players choosing outsider to introduce their 'foreigner' character and promptly dying dead-ly.

Also rename the dept for colonists/visitors from 'contractors' to 'Independants'.

Current issue; can't ready up as visitor for reasons Im not entirely clear on.

* maybe helps

* fixes visitors not being able to add them to preferences