Buffs marshals for Pneumo. Nerfs marshals for Meeeeeee also some bugfixes i guess (#4768)
* Buffs marshals for Pneumo. Nerfs marshals for Meeeeeee

Adds the wind(ostwind + Nordwind) frame to the Marshal disk.

Raises the price of the wind frame to include osmium owing to the high quality of the weapons it produces- ditto for the omni frame. In the future I may want to find an alternative resource for these higher quality weapon frames

* hatever

fixes #4767 probably. Adds a check to prevent bleeding once blood has reached a certain volume.

fixes #4763 , im not sure how this ever worked but should be fine now.

Probably fixes #4740, there was a pathing issue.

First two were all trilby :)

* changes osmium cost to diamond, fixes potential duping issue by adding diamond to guns themselves.

* actually fixes the costs lol