ammo fab changes.
Ammo fab has an additional bypass for high vig(40), and mec requirements down  to 25 from 40. This is intended to have a few effects; for one, it should increase the number of people who could(theoretically)  actually use the ammo fab in an emergency should a WO/Spec not be available. This includes those who might say....break into the armory to help themselves. I also found the 40 req on mech to be a little too high overall, it was another in a long line of 'stat checks that are essentially intended to force you to use the perk bypass'.