Creating ERT Ready Room (#4780)
* Creating ERT Ready Room

Adding the ERT Ready Room.

* Turning Scrap into (what) Matter(s)

This turns the refined scrap into six full stacks of compressed matter. Should do the trick, right?

* Adjusting to Moon's requests.

This changes the room size, additionally it adds vouchers for the Marshals and Corpsman units. 

Adds wardrobes for them, and vending machines where they can get loadouts and even pick up their holsters. If any additional items were required to be vended, anyone with ADMIN+ or SPAWN+ can spawn in some CASH NOW. (Call JG Wentsworth)

* Emergency Floor Fix!!

I realized which floors were what on Centcom and that they actually do up a pattern and are not just the basic tiles from the mapper.

SO TO AVOID ONE EYESORE OF A ROOM. We are YEETING that bitch OUT.

* Nerfing ERT 

Removing the assault armor for SCAF suits for the ERT. 

Guess we can't all be Maxtac Death Machines.