re-adds some guns that were requested (#4781)
* re-adds some guns that were requested

Brings back the bulldog and pitbull and sts40 by popular demand.

STS Para can no longer be crafted- whos idea was this? It's meant to be a relatively rare, custom tooled gun made by/for the shield. It's been added to the leveled lists but i do not think it needs to be craftable.

Adds....the mac10 back in, it's in the leveled lists and has a chance to be in the prospie/salvager locker(replaced the grease gun as no one seemed to use it? we can change this if someone gripes), i only did this because Mojave asked. what a silly gun

Croon is back in the loot pile a bit more rarely than the mac10- it's not good enough to jsutify being in the rare loot pile so it is not.

* stops people from disassembling the mac which haasnt got ugnparts

* reverts ghostroles chjanges

reverts

* Update sts.dm

* Update code/game/objects/items/weapons/autolathe_disk/guns_ammo_sec_disks.dm

Co-authored-by: Trilbyspaceclone <30435998+Trilbyspaceclone@users.noreply.github.com>

* Update code/modules/projectiles/guns/projectile/automatic/pitbull.dm

Co-authored-by: Trilbyspaceclone <30435998+Trilbyspaceclone@users.noreply.github.com>

* Update code/game/objects/items/weapons/autolathe_disk/guns_ammo_sec_disks.dm

Co-authored-by: Trilbyspaceclone <30435998+Trilbyspaceclone@users.noreply.github.com>

* Update code/datums/autolathe/guns.dm

Co-authored-by: Trilbyspaceclone <30435998+Trilbyspaceclone@users.noreply.github.com>

* Update code/modules/projectiles/guns/projectile/automatic/pitbull.dm

Co-authored-by: Trilbyspaceclone <30435998+Trilbyspaceclone@users.noreply.github.com>

---------

Co-authored-by: Trilbyspaceclone <30435998+Trilbyspaceclone@users.noreply.github.com>