Let there be light.
Maintenance now has a number of APCs mapped in, wired but turned off. If you feel up to the task, take to the tunnels like the adepts of old and get that power cranked - or tear the things up when drifters start sucking power from your net! Be careful, however- The wires connecting the APCs are a bit old and exposed wire may have begun to fray exposing mainline voltages to open air.

Also adds a few lights in maintenance in areas that had not had them- presumably due to not having power till now. tsk tsk, bad level design no?