Another megaPR looming on the horizon (#4797)
Adds a new manual for the roboticist(and players) that includes pictures of and a bit of lore for all the various borg frames. - You can either grab it in the loadout or ask the roboticist to see it, incase you were thinking of buying a bot of your own

Adjusts an old noticeboard paper that was incorrect.