end of an era. Some nerfs and buffs
SMGs have increased firerate, generally by 1 step(200rpm->300rpm. 300->600. etc)

Specop nerfed. No longer has SMG lossless full auto, reduced damage by 20% as a base and changes the name to reflect that this is not a big SMG....but rather a small carbine.

Explosive shotgun shells removed from naturally spawning in the rare-ammo pool. Explosive shells cost in the lathe increased by a factor of two.

Incendiaries cost raised by factor of 1.5

Magazines for SBAW that holdPayload raised to roughly be in line with the box. SBAW mag holds about a third as many rounds as a box, so costs about a third as much.

Cost of explosive grenades increased across the board, this includes EMPS, Frags, Blast.

A box of AMR standard cost 1650, but a handful of off-button was 2400...? This has been amended. A handful of AMR rounds(5 bullets) is now 2000, a box is 12,000