more stuff
Removes the laurels bio protection.

Absolutist wintercoat given a bit of stiffness to compensate for having okayish armor(15 melee, energy, bullet.)

Adds a pair of lightly armored pieces. A Jingasa(steel conical hat) and an Ashigaru(Armored apron). They have about as much protection as a leather jacket and are available in the loadout for 1 point each

Adds in several new pieces of headware. 2 straw hats, 4 sunhats, a tiara a headdress and 1 funky helmet

Adds a buncha new undershirts.

Some actual twit put swimsuits in the suit section despite the fact that they're uniforms, they've been moved to under/misc.

Adds a new swimsuit, 1 new gown(with an alt) and an alternate barmaid uniform(directly under the fantasy barmaid dress in the loadout).