Buffs the fenrir a bit.
To compensate for the fact that it is A. rare as hell B. incredibly difficult to use(slowdown, needs to be braced, slow to wield) the Fenrir has been given back its old 1.2 damage value.

Fenrir has also been given 25% less recoil to *try* to make it less sillymode, it's still largely going to require bracing but should now at least be useable

Kills handholding. Guns will no longer complain at you when fired without brace/wielding. Mostly, this is because these messages happening repeatedly was causing lag for some players(I can't confirm if this was due to the messages themselves, or because every single shot ran through an extra set of switches to determine what message to print/roll dice to see if it even prints a message). If people REALLY want, we can add them back later.