ugh
Also adds a new message to warn the player when internal damage is being caused specifically by an infection. hopefuly this will result in fewer people ignoring internal damage warnings that turn out to be by infections that proceed to kill them dead. A

Spaceacillian now works on corpses, so as to make dealing with folks who die via infection less...impossible.