Mech buffs whooo
mech "armor" raised by 1 step. This one is a lil weird so follow with. Rather than having armor represented the way it is for other things(melee=x bullet=x energy=x) it's instead represented via two values. one is 'armor' a number between 1 and 5 and the other absorbtion, a flat reduction to incoming damage.  Armor has been raised from 1-5 to 2-6, so everything gets a step higher armor, this means that a ripley for instance has twice as much armor, but a seraph has only 20% more. Absorbtion values not touched.

This has been paired with a 50% increase in exo health for all combat mechs.

For now, that is all. This will likely need further tweaking especially for big boy mechs that're rarely seen like the Seraph and such, but should go a ways towards making mecha a bit less flimsy in the face of small arms fire.