returns the Watchtower.
Troopers are now offered the Watchtower instead of the Roe(who's idea even was it to give the shield a 6.5 sniper with a scope so far that rounds will be pellets by the time they hit the target??? Never seen this used!!!)

Also adds the long awaited tan cover to match the tan fatigues that we've had for at least 2 years.  Credit to - Baaaay I think? for the sprite.

Also also adds the zhan headscarf as a headscarf under loadout. Some may remember it as a garment liked by some Tajara and many larpers who thought it was shemagh.