A buncha fucking nonsense.
More tweaking for ERTs included.

2 new sets of armor - a RIG and a set of regular armor. The former a slightly less slow shield RIG, the latter essentially equivalent to the mercenary vest/helmet. The armor is ERT only however the WO/CO get a version that has stats closer to their actual dept rigs.

armor/rig are both bay sprites that I'm hapy w/ the quality of. Thank u bay

 ERT fellas now get a speshial ID card with an excellently coder-sprited 'H.C' on it. Instead of every agent getting a special unique ID with varying levels of access and requiring a lotta manual fucking with, instead they're all children of the agent ID and start with AA. This way, you can instantly imprint your details on, change your name on it to whatever your characters name is and boom.

Further finishes mapping the ERT ready room and such. The area includes a teleporter, a guild prep area with supplies(Gee i wonder why this was necessary).

The armory has been completely reworked, instead being mostly specialized powerful security gear and including far fewer redundant pieces - a rack absolutely laden with vouchers is included as well as a blackshield voucher vendor that was....missing.

Militia/Marshal ERT get a new badge that's half silver half gold to go with the fact it's a dual mark that carries the authority of both the Provost and the Brigadier.